Correct error in conversion of two 1-byte integers to a 2-byte integer

+++ AVRTools/Analog2Digital.cpp
@@ -152,5 +152,5 @@ int readA2D( int8_t channel )
         ;
 
     // NOTE: must read ADCL before ADCH
-    return  ADCL | ( ADCH << 8 );
+    return  ADCL | ( static_cast<uint16_t>(ADCH) << 8 );
 }
